requirements-dev.txt
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
factory-boy>=3.3.0
ruff>=0.0.292
pre-commit>=3.5.0
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
